+++ src/components/HeroImage.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import HeroUtil from '../models/HeroUtil'
 
 const knownHeroes = [
-  'ana', 'bastion', 'brigitte', 'doomfist', 'dva', 'genji',
+  'ana', 'ashe', 'bastion', 'brigitte', 'doomfist', 'dva', 'genji',
   'hanzo', 'junkrat', 'lucio', 'mccree', 'mei', 'mercy', 'moira',
   'orisa', 'pharah', 'reaper', 'reinhardt', 'roadhog', 'soldier76',
   'sombra', 'symmetra', 'torbjorn', 'tracer', 'widowmaker', 'winston',