+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added support for a 'joined voice' field when importing from a CSV file, and exporting matches, to specify whether you joined voice chat in a match
 - Added Ashe [#48](https://github.com/cheshire137/competiwatch-desktop/issues/48)
 - Added competitive season 13 [#47](https://github.com/cheshire137/competiwatch-desktop/issues/47)
+- Fixed UI bug where a horizontal scrollbar could appear when hovering over the match edit button on the Matches page
 
 ## [Version 0.1.9](https://github.com/cheshire137/competiwatch-desktop/releases/tag/0.1.9)