+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## Version 0.2.1
+
+- Added chart to Trends tab showing wins and losses by whether or not you joined voice chat
+
 ## Version 0.2.0
 
 - Added 'Copy from last match' to the Group Members field when logging a match [#43](https://github.com/cheshire137/competiwatch-desktop/issues/43)